python-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
asyncpg
playwright
beautifulsoup4